requirements.txt
Flask==3.0.0
numpy==1.26.2
tensorflow-cpu==2.15.0
pandas==2.1.4
Pillow==10.1.0
gunicorn==21.2.0
